+++ packages/components/src/inputs/TextInput.tsx
@@ -1,0 +1,4 @@
+export {
+  default as TextInput,
+  TextFieldProps as TextInputProps,
+} from '@material-ui/core/TextField';